# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GameForge-Copilot",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/games/{game_id}/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_v1_games__game_id__runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RunResp_"
                }
              }
            }
          },
          "404": {
            "description": "run 或游戏不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "邮箱未验证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "配额耗尽",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_v1_games__game_id__runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_RunListItem__"
                }
              }
            }
          },
          "404": {
            "description": "run 或游戏不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RunStatusResp_"
                }
              }
            }
          },
          "404": {
            "description": "run 或游戏不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/pause": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Pause Run",
        "operationId": "pause_run_api_v1_runs__run_id__pause_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RunControlResp_"
                }
              }
            }
          },
          "404": {
            "description": "run 或游戏不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/resume": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Resume Run",
        "operationId": "resume_run_api_v1_runs__run_id__resume_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RunControlResp_"
                }
              }
            }
          },
          "404": {
            "description": "run 或游戏不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Cancel Run",
        "operationId": "cancel_run_api_v1_runs__run_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RunControlResp_"
                }
              }
            }
          },
          "404": {
            "description": "run 或游戏不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/retry": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Retry Run",
        "description": "从失败阶段重试（Batch A · R3）。",
        "operationId": "retry_run_api_v1_runs__run_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RunControlResp_"
                }
              }
            }
          },
          "404": {
            "description": "run 或游戏不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{game_id}/runs/{run_id}/hitl/resolve": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Resolve Hitl",
        "description": "解决 HITL：plan_confirm / sandbox_failed / qa_failed → enqueue resume。",
        "operationId": "resolve_hitl_api_v1_games__game_id__runs__run_id__hitl_resolve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Game Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HitlResolveReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_HitlResolveResp_"
                }
              }
            }
          },
          "404": {
            "description": "run 或游戏不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_HitlResolveResp_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/HitlResolveResp"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[HitlResolveResp]"
      },
...
      "ApiResponse_RunControlResp_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RunControlResp"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[RunControlResp]"
      },
      "ApiResponse_RunResp_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RunResp"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[RunResp]"
      },
      "ApiResponse_RunStatusResp_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RunStatusResp"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[RunStatusResp]"
      },
...
      "ApiResponse_list_RunListItem__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RunListItem"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[list[RunListItem]]"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "统一错误响应：{\"error\": {code, message, detail}}"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HitlResolveReq": {
        "properties": {
          "node": {
            "type": "string",
            "title": "Node"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "modify_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modify Text"
          }
        },
        "type": "object",
        "required": [
          "node",
          "decision"
        ],
        "title": "HitlResolveReq"
      },
      "HitlResolveResp": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "phase": {
            "$ref": "#/components/schemas/RunPhase"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "phase"
        ],
        "title": "HitlResolveResp"
      },
      "HitlState": {
        "properties": {
          "node": {
            "type": "string",
            "title": "Node"
          }
        },
        "type": "object",
        "required": [
          "node"
        ],
        "title": "HitlState"
      },
...
      "RunControlResp": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "phase": {
            "$ref": "#/components/schemas/RunPhase"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "phase"
        ],
        "title": "RunControlResp"
      },
      "RunCreate": {
        "properties": {
          "requirement": {
            "type": "string",
            "title": "Requirement"
          },
          "llm_config_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Config Id"
          }
        },
        "type": "object",
        "required": [
          "requirement"
        ],
        "title": "RunCreate"
      },
      "RunListItem": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "phase": {
            "$ref": "#/components/schemas/RunPhase"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "phase",
          "started_at"
        ],
        "title": "RunListItem"
      },
      "RunPhase": {
        "type": "string",
        "enum": [
          "plan",
          "art",
          "code",
          "qa",
          "done"
        ],
        "title": "RunPhase"
      },
      "RunResp": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "game_id": {
            "type": "string",
            "format": "uuid",
            "title": "Game Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "phase": {
            "$ref": "#/components/schemas/RunPhase"
          },
          "ws_url": {
            "type": "string",
            "title": "Ws Url"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "game_id",
          "status",
          "phase",
          "ws_url"
        ],
        "title": "RunResp"
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "running",
          "paused",
          "done",
          "failed"
        ],
        "title": "RunStatus"
      },
      "RunStatusResp": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "game_id": {
            "type": "string",
            "format": "uuid",
            "title": "Game Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "phase": {
            "$ref": "#/components/schemas/RunPhase"
          },
          "ws_url": {
            "type": "string",
            "title": "Ws Url"
          },
          "current_hitl": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HitlState"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "game_id",
          "status",
          "phase",
          "ws_url"
        ],
        "title": "RunStatusResp"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}